+++ src/Services/ApiPaymentService.php
@@ -5,18 +5,11 @@ namespace Uca\Payments\Services;
 use Illuminate\Support\Facades\Http;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 use Symfony\Component\HttpFoundation\Response as HttpFoundationResponse;
-use Uca\Payments\Data\PaymentGatewayData;
 use Illuminate\Support\Facades\Cache;
-use Illuminate\Http\Client\Response;
 
 class ApiPaymentService
 {
     private const ENDPOINTS = [
-        'external_reference' => 'payment/remote/{paymentGatewayId}/external_reference/{externalReference}',
-        'transaction_id' => 'payment/remote/{paymentGatewayId}/gateway_transaction_id/{transactionId}',
-        'search' => 'payment/local/search?',
-        'payment_gateway' => 'payment-gateway',
-        'payment_sync' => 'payment/{uniqueField}/{value}/sync',
         'auth_token' => 'auth/token',
     ];
 
@@ -36,59 +29,14 @@ class ApiPaymentService
 
         $this->token = Cache::get('payment_gateway_access_token') ?? $this->refreshAccessToken();
     }
-    public function sync(string $unique_field, string $value): array
-    {
-        $url_params = [
-            'uniqueField' => $unique_field,
-            'value' => $value
-        ];
-        return $this->doPut(self::ENDPOINTS['payment_sync'], $url_params);
-    }
 
-    public function byExternalReference(string $payment_gateway_id, string $external_reference): array
+    protected function makeUrlFromBody(array $params)
     {
-        $params = [
-            'externalReference' => $external_reference,
-            'paymentGatewayId' => $payment_gateway_id
-        ];
-        return $this->doGet(self::ENDPOINTS['external_reference'], $params);
-    }
-
-    public function byTransactionId(string $payment_gateway_id, string $transaction_id): array
-    {
-        $params = [
-            'transactionId' => $transaction_id,
-            'paymentGatewayId' => $payment_gateway_id
-        ];
-        return $this->doGet(self::ENDPOINTS['transaction_id'], $params);
-    }
-
-    public function byPreferenceId(string $preference_id): array
-    {
-        $params = ['preference_id' => $preference_id];
-        return $this->search($params);
-    }
-
-    public function byId(string $id): array
-    {
-        $params = ['id' => $id];
-        return $this->search($params);
-    }
-
-    public function search(array $params = []): array
-    {
-        $endpoint = self::ENDPOINTS['search'];
         // Tomamos solo las keys
         $keys = array_keys($params);
         // Genera "/{id}/{status}/{date}"
-        $endpoint .= implode('&', array_map(fn($k) => $k . '={' . $k . '}', $keys));
-        return $this->doGet($endpoint, $params);
-    }
-
-    public function updateOrCreatePaymentGateway(PaymentGatewayData $payment_gateway): array
-    {
-        $response = $this->doPost(self::ENDPOINTS['payment_gateway'], $payment_gateway->toArray());
-        return  $response['data'];
+        $endpoint = implode('&', array_map(fn($k) => $k . '={' . $k . '}', $keys));
+        return $endpoint;
     }
 
     private function refreshAccessToken(): string
@@ -107,7 +55,7 @@ class ApiPaymentService
         throw new \Exception('Authentication failed: ' . $response->body());
     }
 
-    private function doGet(string $endpoint, array $params, bool $retry = true): array
+    protected function doGet(string $endpoint, array $params, bool $retry = true): array
     {
         $response = Http::withToken($this->token)
             ->withHeaders([
@@ -138,7 +86,7 @@ class ApiPaymentService
         }
     }
 
-    private function doPost(string $endpoint, array $params, bool $retry = true): array
+    protected function doPost(string $endpoint, array $params, bool $retry = true): array
     {
         $response = Http::withToken($this->token)
             ->withHeaders([
@@ -159,7 +107,7 @@ class ApiPaymentService
         return  $response->json();
     }
 
-    private function doPut(string $endpoint, array $url_params = [], array $body_params = [], bool $retry = true): array
+    protected function doPut(string $endpoint, array $url_params = [], array $body_params = [], bool $retry = true): array
     {
         $response = Http::withToken($this->token)
             ->withHeaders([